GREEN: Deposit, Withdraw, Overdraft. Updated model

+++ Boolean.CSharp.Main/ExpensesAccount.cs
@@ -8,7 +8,7 @@ namespace Boolean.CSharp.Main
 {
     public class ExpensesAccount : Account
     {
-        public ExpensesAccount(float branchcode, string customerPhoneNumber) : base(branchcode, customerPhoneNumber)
+        public ExpensesAccount(int branchcode, string customerPhoneNumber) : base(branchcode, customerPhoneNumber)
         {
         }
     }
